Add files via upload
This project predicts the prices of laptops based on various features using machine learning models. The dataset includes specifications such as brand, processor type, RAM, storage, GPU, operating system, and more. The goal is to build a regression model that can accurately estimate the price of a laptop based on its configuration.